2005-01-24	Ralf Corsepius <ralf.corsepius@rtems.org>
	* score/include/rtems/score/heap.h: Use uintptr_t for _H_uptr_t.
	* score/inline/rtems/score/address.inl: Remove
	RTEMS_CPU_HAS_16_BIT_ADDRESSES.
	Use uintptr_t for void* to address cast.

+++ cpukit/score/include/rtems/score/heap.h
@@ -51,7 +51,7 @@ extern "C" {
  *        'rtems/score/types.h' and used here.
  */
 
-typedef unsigned long int _H_uptr_t;
+typedef uintptr_t _H_uptr_t;
 
 /**
  *  Forward reference
